Backend/requirements.txt
cryptography==3.4.8
Flask==3.0.0
Flask_JWT_Extended==4.5.2
flask_socketio==5.3.6
mysql_connector_repackaged==0.3.1
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
PyJWT==2.3.0
PyJWT==2.8.0
PyPDF2==3.0.1
python-dotenv==1.0.0
tabula==1.0.5
